refactor: move CodeEditor component to the editor directory and update all import paths
