--- components/CodeEditor/index.ts
@@ -1,1 +1,0 @@
-export { CodeEditor, type CodeEditorProps, type CodeEditorLanguage } from "./CodeEditor";
