+++ components/problems/ProblemUpsertForm.tsx
@@ -19,7 +19,7 @@ import {
   loadQuestionCategoryOptionsForForm,
   loadTagOptionsForForm,
 } from "@/lib/questionTaxonomyApi";
-import CodeEditor from "../CodeEditor/CodeEditor";
+import CodeEditor from "../editor/CodeEditor";
 
 type ProblemDetailResponse = {
   category_id?: string | number | null;
